feat: Move state publishing configuration to Clockwork from NATS in conf file
